+++ extension.js
@@ -2,6 +2,8 @@ const vscode = require("vscode");
 const languages = require("./languages.js");
 const translate = require("google-translate-open-api").default;
 const he = require("he");
+const humanizeString = require("humanize-string");
+const camelcase = require("camelcase");
 
 /**
  * @typedef TranslateRes
@@ -23,7 +25,7 @@ const recentlyUsed = [];
  * @returns {undefined}
  */
 function updateLanguageList(selectedLanguage) {
-  const index = recentlyUsed.findIndex(r => r === selectedLanguage);
+  const index = recentlyUsed.findIndex((r) => r === selectedLanguage);
   if (index !== -1) {
     // Remove the recently used language from the list
     recentlyUsed.splice(index, 1);
@@ -80,28 +82,47 @@ function getTranslationPromise(selectedText, selectedLanguage, selection) {
       translationConfiguration.proxy = {
         host,
         port: Number(port),
-      }
+      };
       if (!!username) {
         translationConfiguration.proxy.auth = {
           username,
           password,
-        }
+        };
       }
     }
     translate(selectedText, translationConfiguration)
-      .then(res => {
+      .then((res) => {
         if (!!res && !!res.data) {
-          resolve(
-            /** @type {TranslateRes} */ {
-              selection,
-              translation: res.data[0]
-            }
-          );
+          // If google rejects the string it will return the same string as input
+          // We can try to split the string into parts, then translate again. Then return it to a
+          // camel style casing
+          if (res.data[0] === selectedText) {
+            translate(humanizeString(selectedText), translationConfiguration)
+              .then(res => {
+                if (!!res && !!res.data) {
+                  resolve(
+                    /** @type {TranslateRes} */ {
+                      selection,
+                      translation: camelcase(res.data[0])
+                    }
+                  );
+                } else {
+                  reject(new Error("Google Translation API issue"));
+                }
+              });
+          } else {
+            resolve(
+              /** @type {TranslateRes} */ {
+                selection,
+                translation: res.data[0]
+              }
+            );
+          }
         } else {
           reject(new Error("Google Translation API issue"));
         }
       })
-      .catch(e =>
+      .catch((e) =>
         reject(new Error("Google Translation API issue: " + e.message))
       );
   });
@@ -116,7 +137,7 @@ function getTranslationPromise(selectedText, selectedLanguage, selection) {
  * @returns {Array.<Promise<TranslateRes>>}
  */
 function getTranslationsPromiseArray(selections, document, selectedLanguage) {
-  return selections.map(selection => {
+  return selections.map((selection) => {
     const selectedText = getSelectedText(document, selection);
     return getTranslationPromise(selectedText, selectedLanguage, selection);
   });
@@ -135,7 +156,7 @@ function getTranslationsPromiseArrayLine(
   document,
   selectedLanguage
 ) {
-  return selections.map(selection => {
+  return selections.map((selection) => {
     const selectedLineText = getSelectedLineText(document, selection);
     return getTranslationPromise(selectedLineText, selectedLanguage, selection);
   });
@@ -145,32 +166,37 @@ function getTranslationsPromiseArrayLine(
  * Returns user settings Preferred language.
  * If user hasn't set preferred lang. Prompt to set.
  */
-async function getPreferredLanguage() {
-  return vscode.workspace
-    .getConfiguration("vscodeGoogleTranslate")
-    .get("preferredLanguage")
-    || setPreferredLanguage();
+function getPreferredLanguage() {
+  return (
+    vscode.workspace
+      .getConfiguration("vscodeGoogleTranslate")
+      .get("preferredLanguage") || setPreferredLanguage()
+  );
 }
 
 function setPreferredLanguage() {
   const editor = vscode.window.activeTextEditor;
 
   const quickPickData = recentlyUsed
-    .map(r => ({
+    .map((r) => ({
       label: r,
       description: "(recently used)",
     }))
-    .concat(languages.map(r => ({ label: r.name })));
+    .concat(languages.map((r) => ({ label: r.name })));
 
-  return vscode.window
-    .showQuickPick(quickPickData)
-    .then(selectedLanguage => {
-      if (!selectedLanguage) {
-        return;
-      }
-      vscode.workspace.getConfiguration().update("vscodeGoogleTranslate.preferredLanguage", selectedLanguage.label, vscode.ConfigurationTarget.Global);
-      return selectedLanguage.label;
-    })
+  return vscode.window.showQuickPick(quickPickData).then((selectedLanguage) => {
+    if (!selectedLanguage) {
+      return;
+    }
+    vscode.workspace
+      .getConfiguration()
+      .update(
+        "vscodeGoogleTranslate.preferredLanguage",
+        selectedLanguage.label,
+        vscode.ConfigurationTarget.Global
+      );
+    return selectedLanguage.label;
+  });
 }
 
 /**
@@ -185,7 +211,7 @@ function getProxyConfig() {
     port: config.get("proxyPort"),
     username: config.get("proxyUsername"),
     password: config.get("proxyPassword"),
-  }
+  };
 }
 
 /**
@@ -195,60 +221,65 @@ function getProxyConfig() {
  * @returns {undefined} There is no an API public surface now (7/3/2019)
  */
 function activate(context) {
-  let translateText = vscode.commands.registerCommand(
+  const translateText = vscode.commands.registerCommand(
     "extension.translateText",
-    function() {
+    function () {
       const editor = vscode.window.activeTextEditor;
       const { document, selections } = editor;
 
       const quickPickData = recentlyUsed
-        .map(r => ({
+        .map((r) => ({
           label: r,
           description: "(recently used)",
         }))
-        .concat(languages.map(r => ({ label: r.name })));
+        .concat(languages.map((r) => ({ label: r.name })));
 
       vscode.window
         .showQuickPick(quickPickData)
-        .then(selectedLanguage => {
+        .then((selectedLanguage) => {
           if (!selectedLanguage) return;
           updateLanguageList(selectedLanguage.label);
           const translationsPromiseArray = getTranslationsPromiseArray(
             selections,
             document,
-            languages.find(r => r.name === selectedLanguage.label).value
+            languages.find((r) => r.name === selectedLanguage.label).value
           );
           Promise.all(translationsPromiseArray)
-            .then(function(results) {
-              editor.edit(builder => {
-                results.forEach(r => {
+            .then(function (results) {
+              editor.edit((builder) => {
+                results.forEach((r) => {
                   if (!!r.translation) {
                     builder.replace(r.selection, he.decode(r.translation));
                   }
                 });
               });
             })
-            .catch(e => vscode.window.showErrorMessage(e.message));
+            .catch((e) => vscode.window.showErrorMessage(e.message));
         })
-        .catch(err => {
+        .catch((err) => {
           vscode.window.showErrorMessage(err.message);
         });
     }
   );
   context.subscriptions.push(translateText);
 
-  let setPreferredLanguageFnc = vscode.commands.registerCommand("extension.setPreferredLanguage", setPreferredLanguage);
+  const setPreferredLanguageFnc = vscode.commands.registerCommand(
+    "extension.setPreferredLanguage",
+    setPreferredLanguage
+  );
   context.subscriptions.push(setPreferredLanguageFnc);
 
-  let translateTextPreferred = vscode.commands.registerCommand(
+  const translateTextPreferred = vscode.commands.registerCommand(
     "extension.translateTextPreferred",
-    async function() {
+    async function () {
       const editor = vscode.window.activeTextEditor;
       const { document, selections } = editor;
 
       // vscodeTranslate.preferredLanguage
-      let preferredLanguage = await getPreferredLanguage();
-      let locale = languages.find(element => element.name === preferredLanguage).value;
+      const preferredLanguage = await getPreferredLanguage();
+      const locale = languages.find(
+        (element) => element.name === preferredLanguage
+      ).value;
       if (!locale) {
         return;
       }
@@ -259,47 +290,47 @@ function activate(context) {
         locale
       );
       Promise.all(translationsPromiseArray)
-        .then(function(results) {
-          editor.edit(builder => {
-            results.forEach(r => {
+        .then(function (results) {
+          editor.edit((builder) => {
+            results.forEach((r) => {
               if (!!r.translation) {
                 builder.replace(r.selection, he.decode(r.translation));
               }
             });
           });
         })
-        .catch(e => vscode.window.showErrorMessage(e.message));
+        .catch((e) => vscode.window.showErrorMessage(e.message));
     }
   );
   context.subscriptions.push(translateTextPreferred);
 
-  let translateLinesUnderCursor = vscode.commands.registerCommand(
+  const translateLinesUnderCursor = vscode.commands.registerCommand(
     "extension.translateLinesUnderCursor",
     function translateLinesUnderCursorcallback() {
       const editor = vscode.window.activeTextEditor;
       const { document, selections } = editor;
 
       const quickPickData = recentlyUsed
-        .map(r => ({
+        .map((r) => ({
           label: r.name,
           description: "(recently used)",
         }))
-        .concat(languages.map(r => ({ label: r.name })));
+        .concat(languages.map((r) => ({ label: r.name })));
 
       vscode.window
         .showQuickPick(quickPickData)
-        .then(selectedLanguage => {
+        .then((selectedLanguage) => {
           if (!selectedLanguage) return;
           updateLanguageList(selectedLanguage.label);
           const translationsPromiseArray = getTranslationsPromiseArrayLine(
             selections,
             document,
-            languages.find(r => r.name === selectedLanguage.label).value
+            languages.find((r) => r.name === selectedLanguage.label).value
           );
           Promise.all(translationsPromiseArray)
-            .then(function(results) {
-              editor.edit(builder => {
-                results.forEach(r => {
+            .then(function (results) {
+              editor.edit((builder) => {
+                results.forEach((r) => {
                   if (!!r.translation) {
                     const ffix = ["", "\n"];
                     if (
@@ -313,9 +344,9 @@ function activate(context) {
                 });
               });
             })
-            .catch(e => vscode.window.showErrorMessage(e.message));
+            .catch((e) => vscode.window.showErrorMessage(e.message));
         })
-        .catch(err => {
+        .catch((err) => {
           vscode.window.showErrorMessage(err.message);
         });
     }
@@ -323,13 +354,15 @@ function activate(context) {
 
   context.subscriptions.push(translateLinesUnderCursor);
 
-  let translateLinesUnderCursorPreferred = vscode.commands.registerCommand(
+  const translateLinesUnderCursorPreferred = vscode.commands.registerCommand(
     "extension.translateLinesUnderCursorPreferred",
     async function translateLinesUnderCursorPreferredcallback() {
       const editor = vscode.window.activeTextEditor;
       const { document, selections } = editor;
-      let preferredLanguage = await getPreferredLanguage();
-      let locale = languages.find(element => element.name === preferredLanguage).value;
+      const preferredLanguage = await getPreferredLanguage();
+      const locale = languages.find(
+        (element) => element.name === preferredLanguage
+      ).value;
       if (!locale) {
         vscode.window.showWarningMessage(
           "Prefered language is requeried for this feature! Please set this in the settings."
@@ -344,9 +377,9 @@ function activate(context) {
       );
 
       Promise.all(translationsPromiseArray)
-        .then(function(results) {
-          editor.edit(builder => {
-            results.forEach(r => {
+        .then(function (results) {
+          editor.edit((builder) => {
+            results.forEach((r) => {
               if (!!r.translation) {
                 const ffix = ["", "\n"];
                 if (editor.document.lineCount - 1 === r.selection.start.line)
@@ -357,7 +390,7 @@ function activate(context) {
             });
           });
         })
-        .catch(e => vscode.window.showErrorMessage(e.message));
+        .catch((e) => vscode.window.showErrorMessage(e.message));
     }
   );