+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 1.4.7
+- Translate camel case support
+
+## 1.4.6
+- Improve default language settings
+- Fix default shortcut
+
 ## 1.4.5
 - Update translation library